Implement smooth time updates
