+++ Sources/Castor/TimeProperties.swift
@@ -1,0 +1,12 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import CoreMedia
+
+struct TimeProperties {
+    let time: CMTime
+    let timeRange: CMTimeRange
+}
